vmm, hypervisor: Feature guard KVM specific code
There are some code bas and function which are purely KVM specific for
now and we don't have those supports in mshv for now but we have plan
for future. We are doing a feature guard with kvm. For example, KVM has
mp_state, cpu clock support,  which we don't have for mshv. In order build
those code we are making the code base for KVM specific compilation.

Signed-off-by: Muminul Islam <muislam@microsoft.com>